+++ src/index.js
@@ -1,6 +1,7 @@
 import mainfunc from './main'
 import rule from './data/rule'
 import environment from './data/environment'
+import worker from './data/worker'
 import setData from './option/setData'
 import notice from './option/noticeData'
 
@@ -11,6 +12,18 @@ mainfunc.install = function(Vue, options = {}) {
   Vue.prototype._func = this
 }
 
-export { rule, environment, notice }
+worker.set({
+  func: function(list) {
+    list.push('111')
+    return list
+  },
+  self: null,
+  args: [[1, 23]]
+}).then(
+  res => { console.log(res) },
+  err => { console.log(err) }
+)
+
+export { rule, environment, notice, worker }
 
 export default mainfunc